+++ Arcade Arena.Library/Animation.cs
@@ -3,10 +3,10 @@ namespace Arcade_Arena.Library
 {
     public class Animation
     {
-        public int XRecPos { get; set; }
-        public int YRecPos { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public short XRecPos { get; set; }
+        public short YRecPos { get; set; }
+        public short Width { get; set; }
+        public short Height { get; set; }
 
         public Animation()
         {
