+++ Arcade Arena.Server/Commands/InputCommand.cs
@@ -19,12 +19,12 @@ namespace Arcade_Arena.Server
                 managerLogger.AddLogMessage("Server", string.Format("Didn't find player with name {0}", name));
                 return;
             }
-            playerAndConnection.Player.XPosition = inc.ReadInt32();
-            playerAndConnection.Player.YPosition = inc.ReadInt32();
-            playerAndConnection.Player.Animation.XRecPos = inc.ReadInt32();
-            playerAndConnection.Player.Animation.YRecPos = inc.ReadInt32();
-            playerAndConnection.Player.Animation.Height = inc.ReadInt32();
-            playerAndConnection.Player.Animation.Width = inc.ReadInt32();
+            playerAndConnection.Player.XPosition = inc.ReadInt16();
+            playerAndConnection.Player.YPosition = inc.ReadInt16();
+            playerAndConnection.Player.Animation.XRecPos = inc.ReadInt16();
+            playerAndConnection.Player.Animation.YRecPos = inc.ReadInt16();
+            playerAndConnection.Player.Animation.Height = inc.ReadInt16();
+            playerAndConnection.Player.Animation.Width = inc.ReadInt16();
             playerAndConnection.Player.Health = inc.ReadSByte();
             playerAndConnection.Player.IntersectingLava = inc.ReadBoolean();
